Fixed signed/unsigned mismatch.

+++ tests/SOCK_Send_Recv_Test.cpp
@@ -1,6 +1,6 @@
 // $Id$
 
-// ============================================================================
+// ===========================================================================
 //
 // = LIBRARY
 //    tests
@@ -74,7 +74,7 @@ client (void *arg)
   // there is no more and no less.
 
   unsigned char buff[255];
-  unsigned int i;
+  size_t  i;
   ssize_t len;
 
   // The server will verify that this data pattern gets there intact.
